requirements.txt
# Sistema Automatizador de Diplomas com QR Code
# Desenvolvido por: Carlos Antonio de Oliveira Piquet
# Email: carlospiquet.projetos@gmail.com
# Contato: +55 21 977434614
# Sistema integrado com Moodle para automação de diplomas

## Dependências Python

Flask==2.3.3
qrcode[pil]==7.4.2
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
MarkupSafe==2.1.3

# PDF processing - alternativa ao PyMuPDF
pypdf2==3.0.1
reportlab==4.0.4
python-magic-bin==0.4.14

# Banco de dados (opcional)
mysql-connector-python==8.1.0
psycopg2-binary==2.9.7

# Desenvolvimento e Testes
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
